+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   and `cider-disconnected-hook`.
 * [#920](https://github.com/clojure-emacs/cider/issues/920): Support `cider-jack-in` for boot-based projects.
 * [#949](https://github.com/clojure-emacs/cider/issues/949): New custom var: `cider-default-repl-command`.
+* New code formatting commands - `cider-format-buffer`, `cider-format-region` and `cider-format-defun`.
 
 ### Changes